pm: switch to dpkg native triggers
As we only care about dpkg now, drop AB3 triggers (which is modelled
after dpkg triggers) and use dpkg one.

In this case, we no longer need to have all packages that will trigger a
package to be in the package's BUILDDEP, because dpkg can do this based
on its DB. In addition, DPKG can postpone trigger actions to the end of
a dpkg/apt command, saves some (much?) time.

AB3 triggers will only react when /var/ab/triggered/XXX is exist, and
new packages packed with dpkg native trigger won't have them, so the
transition shouldn't be problematic.

Signed-off-by: Icenowy Zheng <icenowy@aosc.io>