expand one-liners
The one-liners are causing false warnings when there is actually no
issue.

    [[ $condition ]] && exec_extra_step

The code above is used as a shortcut of a if branch throughout AB3:

    if [[ $condition ]]; then
        exec_extra_step
    fi

The short one-liners cause a warning to be printed out when $condition
evaluates to false as the entire expression evaluates to a non-zero
return-node.

This expansion also allows AB3 to generate a warning when AB3 failed to
find a copyright license file to install.